[project] create separate artifact for anko and remove initial dependency on it. Split project into two artifacts.
