+++ app/src/main/java/com/andrewgrosner/kbinding/sample/MainActivity.kt
@@ -1,0 +1,43 @@
+package com.andrewgrosner.kbinding.sample
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import android.view.Menu
+import android.view.MenuItem
+import org.jetbrains.anko.setContentView
+
+class MainActivity : AppCompatActivity() {
+
+    var layout: MainActivityLayout2? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val viewModel = MainActivityViewModel()
+        layout = MainActivityLayout2(viewModel).apply { setContentView(this@MainActivity) }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        val id = item.itemId
+
+
+        if (id == R.id.action_settings) {
+            return true
+        }
+
+        return super.onOptionsItemSelected(item)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        layout?.destroyView()
+    }
+}
