+++ app/src/main/java/com/andrewgrosner/kbinding/sample/MainActivityLayout2.kt
@@ -1,0 +1,104 @@
+package com.andrewgrosner.kbinding.sample
+
+import android.view.ViewGroup.LayoutParams.MATCH_PARENT
+import com.andrewgrosner.kbinding.anko.BindingComponent
+import com.andrewgrosner.kbinding.anko.bind
+import com.andrewgrosner.kbinding.bindings.*
+import org.jetbrains.anko.*
+import java.util.Calendar.*
+
+/**
+ * Description:
+ */
+class MainActivityLayout2(mainActivityViewModel: MainActivityViewModel)
+    : BindingComponent<MainActivity, MainActivityViewModel>(mainActivityViewModel) {
+
+    val nonNonNullViewModel = mainActivityViewModel
+
+    override fun createViewWithBindings(ui: AnkoContext<MainActivity>) = with(ui) {
+        scrollView {
+            verticalLayout {
+                linearLayout {
+                    textView {
+                        id = R.id.firstName
+                        textSize = 16f
+                        bindSelf(MainActivityViewModel::firstName) { it.firstName }.toText(this)
+                        bind(MainActivityViewModel::firstName) { it.firstName }
+                                .onIsNotNullOrEmpty()
+                                .toViewVisibilityB(this)
+
+                        onClick { viewModel?.onFirstNameClick() }
+                    }.lparams {
+                        rightMargin = dip(4)
+                    }
+
+                    textView {
+                        textSize = 16f
+                        bindSelf(MainActivityViewModel::lastName) { it.lastName }.toText(this)
+                        onClick { viewModel?.onLastNameClick() }
+                    }.lparams {
+                        leftMargin = dip(4)
+                    }
+                }.lparams {
+                    width = MATCH_PARENT
+                }
+
+                editText {
+                    hint = "Text mirrors below (Two way)"
+                    bindSelf { it.formInput }.toText(this)
+                            .twoWay().toFieldFromText()
+                }.lparams {
+                    width = MATCH_PARENT
+                }
+
+                textView {
+                    bindSelf { it.formInput }.toText(this)
+                }
+
+                editText {
+                    hint = "Text mirrors below (One way to source)"
+                    bind(this).onSelf()
+                            .toObservable { it.oneWaySourceInput }
+                }
+
+                textView {
+                    bindSelf { it.oneWaySourceInput }.toText(this)
+                }
+
+                switch {
+                    bindSelf { it.selected }.toOnCheckedChange(this)
+                            .twoWay().toFieldFromCompound().onExpression { _, input ->
+                        text = if (input ?: false) "On" else "Off"
+                    }
+                }
+
+                textView {
+                    bindSelf { it.currentTime }.toView(this) { _, value ->
+                        text = "Current Date is ${value?.get(MONTH)}/${value?.get(DAY_OF_MONTH)}/${value?.get(YEAR)}"
+                    }
+                }
+
+                datePicker {
+                    bindSelf { it.currentTime }.toDatePicker(this)
+                            .twoWay().toFieldFromDate()
+                }
+
+                button {
+                    bindNullable { it }.on { if (it == null) "Set Not Null" else "Set Null" }
+                            .toText(this)
+                    onClick {
+                        if (viewModel != null) {
+                            viewModel = null
+                        } else {
+                            viewModel = nonNonNullViewModel
+                        }
+                    }
+                }
+            }.lparams {
+                horizontalMargin = dip(8)
+            }
+        }
+    }
+
+}
+
