+++ gradle.properties
@@ -7,4 +7,3 @@ bt_licenseName=The MIT License
 bt_licenseUrl=http://opensource.org/licenses/MIT
 bt_repo=Libraries
 bt_packaging=jar
-bt_name=kbinding
