+++ kbinding-anko/gradle.properties
@@ -1,0 +1,1 @@
+bt_name=kbinding-anko
