+++ kbinding-anko/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="com.andrewgrosner.kbinding.anko">
+
+    <application/>
+
+</manifest>
