+++ kbinding/src/main/AndroidManifest.xml
@@ -1,0 +1,13 @@
+<manifest package="com.andrewgrosner.kbinding"
+
+          xmlns:android="http://schemas.android.com/apk/res/android"
+>
+
+    <application android:allowBackup="true"
+                 android:label="@string/app_name"
+                 android:supportsRtl="true"
+    >
+
+    </application>
+
+</manifest>
