+++ kbinding/src/main/java/com/andrewgrosner/kbinding/viewextensions/CompoundButtonExtensions.kt
@@ -1,0 +1,10 @@
+package com.andrewgrosner.kbinding.viewextensions
+
+import android.widget.CompoundButton
+
+
+fun CompoundButton.setCheckedIfNecessary(checked: Boolean?) {
+    if (checked != null && isChecked != checked) {
+        isChecked = checked
+    }
+}
