+++ kbinding/src/main/java/com/andrewgrosner/kbinding/viewextensions/ProgressBarExtensions.kt
@@ -1,0 +1,9 @@
+package com.andrewgrosner.kbinding.viewextensions
+
+import android.widget.ProgressBar
+
+fun ProgressBar.setProgressIfNecessary(progress: Int?) {
+    if (this.progress != progress) {
+        this.progress = progress ?: 0
+    }
+}
