+++ kbinding/src/main/res/values/strings.xml
