--- okbinding/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<manifest package="com.andrewgrosner.okbinding"
-
-          xmlns:android="http://schemas.android.com/apk/res/android"
->
-
-    <application android:allowBackup="true"
-                 android:label="@string/app_name"
-                 android:supportsRtl="true"
-    >
-
-    </application>
-
-</manifest>
