--- okbinding/src/main/java/com/andrewgrosner/okbinding/viewextensions/CompoundButtonExtensions.kt
@@ -1,10 +1,0 @@
-package com.andrewgrosner.okbinding.viewextensions
-
-import android.widget.CompoundButton
-
-
-fun CompoundButton.setCheckedIfNecessary(checked: Boolean?) {
-    if (checked != null && isChecked != checked) {
-        isChecked = checked
-    }
-}
