--- okbinding/src/main/java/com/andrewgrosner/okbinding/viewextensions/ProgressBarExtensions.kt
@@ -1,9 +1,0 @@
-package com.andrewgrosner.okbinding.viewextensions
-
-import android.widget.ProgressBar
-
-fun ProgressBar.setProgressIfNecessary(progress: Int?) {
-    if (this.progress != progress) {
-        this.progress = progress ?: 0
-    }
-}
