--- okbinding/src/main/java/com/andrewgrosner/okbinding/viewextensions/RatingBarExtensions.kt
@@ -1,9 +1,0 @@
-package com.andrewgrosner.okbinding.viewextensions
-
-import android.widget.RatingBar
-
-fun RatingBar.setRatingIfNecessary(rating: Float?) {
-    if (this.rating != rating) {
-        this.rating = rating ?: 0.0f
-    }
-}
