--- okbinding/src/test/java/com/andrewgrosner/okbinding/BindingTests.kt
@@ -1,10 +1,0 @@
-package com.andrewgrosner.okbinding
-
-/**
- * Description:
- */
-class BindingTests {
-
-
-
-}
